requirements.txt
asgiref==3.2.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1
django-appconf==1.0.5
django-autocomplete-light==3.9.1
django-environ==0.8.1
django-heroku==0.3.1
django-import-export==2.7.1
django-storages==1.12.3
django-widget-tweaks==1.4.12
dropbox==11.27.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
ply==3.11
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
tablib==3.2.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
